Demo wiring for graphql subscriptions in container component.
